+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `cli/cli`: Add program name before error messages
 - `Makefile`: Add release build
 - `lib/col`: Show list of project columns
+- `tman`: Add command to rename project
 
 ### Changed